+++ src/hamiltonian.jl
@@ -67,6 +67,18 @@ function ∂H∂r(h::Hamiltonian{<:DenseEuclideanMetric,<:GaussianKinetic}, r::A
     return M⁻¹ * r
 end
 
+function ∂H∂r(
+    h::Hamiltonian{<:RankUpdateEuclideanMetric,<:GaussianKinetic}, r::AbstractVecOrMat
+)
+    (; A, B, D) = h.metric
+    axes_A = __axes(A)
+    axes_r = __axes(r)
+    (last(axes_A) != first(axes_r)) && throw(
+        ArgumentError(lazy"AxesMismatch: A has axes $(axes_A) but r has axes $(axes_r)")
+    )
+    return muladd(A, r, B * (D * (B' * r)))
+end
+
 # TODO (kai) make the order of θ and r consistent with neg_energy
 # TODO (kai) add stricter types to block hamiltonian.jl#L37 from working on unknown metric/kinetic
 # The gradient of a position-dependent Hamiltonian system depends on both θ and r. 
@@ -173,6 +185,14 @@ function neg_energy(
 ) where {T<:AbstractVecOrMat}
     mul!(h.metric._temp, h.metric.M⁻¹, r)
     return -dot(r, h.metric._temp) / 2
+end
+
+function neg_energy(
+    h::Hamiltonian{<:RankUpdateEuclideanMetric,<:GaussianKinetic}, r::T, θ::T
+) where {T<:AbstractVector}
+    (; A, B, D) = h.metric
+    Btr = B' * r
+    return -(dot(r, A, r) + _dot(Btr, D, Btr)) / 2
 end
 
 energy(args...) = -neg_energy(args...)